Refactor method names, to allow for growth into future channel/participant types
